feat: moved supplementary code into helper folder
