+++ helpers/logging.py
@@ -1,0 +1,53 @@
+from datetime import datetime, timezone
+
+from communities.models import Community
+from institutions.models import Institution
+from researchers.models import Researcher
+
+
+def get_current_timestamp():
+    return datetime.now(timezone.utc).strftime("%Y-%m-%d %H:%M:%S")
+
+
+def get_log_data(instance):
+    log_data = {
+        'UTC-TIME': get_current_timestamp(),
+        'RELATED-RESEARCHER': None,
+        'RELATED-COMMUNITIES': [],
+        'RELATED-INSTITUTIONS': [],
+    }
+
+    # get related researcher data of interest for this user
+    if Researcher.objects.filter(user=instance).exists():
+        researcher = Researcher.objects.get(user=instance)
+        log_data['RELATED-RESEARCHER'] = {
+            'id': researcher.id,
+            'orcid': researcher.orcid,
+            'contact_email': researcher.contact_email,
+        }
+
+    # get related community data of interest for this user
+    for community in Community.objects.filter(community_creator=instance):
+        log_data['RELATED-COMMUNITIES'].append(
+            {
+                'id': community.id,
+                'community_name': community.community_name,
+                'contact_email': community.contact_email,
+                'is_approved': community.is_approved,
+
+            }
+        )
+
+    # get related institution data of interest for this user
+    for institution in Institution.objects.filter(institution_creator=instance):
+        log_data['RELATED-INSTITUTIONS'].append(
+            {
+                'id': institution.id,
+                'institution_name': institution.institution_name,
+                'institution_email': institution.contact_email,
+                'is_approved': institution.is_approved,
+
+            }
+        )
+
+    return log_data
