+++ tests/unit/test_log_deleted_user.py
@@ -1,16 +1,18 @@
 from unittest.mock import patch
 
+import pytest
 from django.test import TransactionTestCase
 from factories.accounts_factories import UserFactory
 from factories.researchers_factories import ResearcherFactory
 from factories.communities_factories import CommunityFactory
 from factories.institutions_factories import InstitutionFactory
 
-from accounts.signals import get_log_data
+from helpers.logging import get_log_data
 
 
 class TestLogDeletedUser(TransactionTestCase):
 
+    @pytest.mark.django_db
     def setUp(self):
         self.user = UserFactory()
         self.researcher = ResearcherFactory(user=self.user)
@@ -26,7 +28,7 @@ class TestLogDeletedUser(TransactionTestCase):
         ]
 
     def test_log_deleted_user(self):
-        with patch('accounts.signals.get_log_data_timestamp') as timestamp_mock:
+        with patch('helpers.logging.get_current_timestamp') as timestamp_mock:
             timestamp_mock.return_value = 'time-place_holder'
 
             with patch('pprint.pprint') as pprint_mock:
